+++ 2P-UnityEscapeGame/Assets/Scripts/Scripts_kbh/arrowScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e1459f7e8756d3449a53f497a2a4c065
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
