+++ 2P-UnityEscapeGame/Assets/Scripts/Scripts_kbh/gameManager3.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: aebbffc4e4fbfb642bf66202847d229d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
