+++ 2P-UnityEscapeGame/Assets/Scripts/Scripts_kbh/talkManager.cs
@@ -6,16 +6,21 @@ using UnityEngine.UI;
 public class talkManager : MonoBehaviour
 {
     Dictionary<int, string[]> talkData;
-
+    public int length;
+    public int value;//npc의 고유번호
+    gameManager3 v;
     void Awake()
     {
         talkData = new Dictionary<int, string[]>();
+        
         GenerateData();
+        CheckLength(value);
     }
-
-    void Update()
+    private void Start()
     {
-
+    }
+    private void Update()
+    {
     }
 
     void GenerateData()
@@ -23,7 +28,7 @@ public class talkManager : MonoBehaviour
         //index, string
         //1.양 (가장 처음 만나는 npc)
         talkData.Add(1, new string[]
-        { "..", "넌 누구야? 여긴 어쩐일이야?","노란 강아지? \n아니, 못봤어. 여기는 양들만 지내는 곳이야."});
+        {  "넌 누구야? 여긴 어쩐일이야?","노란 강아지? \n아니, 못봤어. 여기는 양들만 지내는 곳이야."});
 
 
 
@@ -32,5 +37,13 @@ public class talkManager : MonoBehaviour
     public string GetTalk(int id, int talkIndex)
     {
         return talkData[id][talkIndex];
+    }
+
+    public int CheckLength(int value)
+    {
+        value = GameObject.Find("Man").GetComponent<gameManager3>().value;
+
+        return length = talkData[value].Length;
+
     }
 }
